pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-search"
version = "0.1.0"
description = "Obsidian block-level relation indexing and search service"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "python-frontmatter>=1.1.0",
  "openai>=1.45.0",
  "tiktoken>=0.7.0",
  "pymilvus>=2.5.0",
  "milvus-lite>=2.5.1",
  "PyYAML>=6.0.2",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
